+++ src/Common.h
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (C) 2026 Ali Yaakub
 // Markdown Peek - shared declarations.
 #pragma once